--- Sources/Custom/CustomTab/CustomTab.swift
@@ -1,9 +1,0 @@
-//
-//  CustomTab.swift
-//
-//  Created by Joe Pan on 2024/10/24.
-//
-
-import UIKit
-
-enum CustomTab {}
