+++ Sources/Custom/CustomTabBarVC/CustomTabBarVC+Model.swift
@@ -1,5 +1,5 @@
 //
-//  CustomTabModel.swift
+//  CustomTabBarVC+Model.swift
 //
 //  Created by Joe Pan on 2024/10/24.
 //
@@ -7,7 +7,7 @@
 
 import UIKit
 
-extension CustomTab {
+extension CustomTabBarVC {
     protocol ScrollToTopable: UIViewController {
         func scrollToTop()
     }
